Code Improvements
Made simple refactoring in the code, to make it simpler and more
maintainable.
With a list of songs, and only a function that receives a key (the name
of the song defined in the button).

Important: The name defined on the button should be the same in the list
of songs. ;)